factor out buildJamboCLI() from cli.js
This commit factors out the buildJamboCLI() method
from cli.js. No functional changes have been made.
This method will be used by the upcoming acceptance test
infrastructure, to create test jambo instances.

J=SLAP-992
TEST=manual

test that I can still build pages with jambo
test that jambo still logs 'You must provide Jambo with a command.'
if you don't enter a command
test that custom commands still work
